url and email extraction for ipa
